+++ packages/pob/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [7.10.0](https://github.com/christophehurpeau/pob/compare/pob@7.9.0...pob@7.10.0) (2021-03-13)
+
+
+### Bug Fixes
+
+* **check-package-dependencies:** better duplicate message error and fix reportError pkgPath ([fede23a](https://github.com/christophehurpeau/pob/commit/fede23a9e417b4e5cadc5dc81baae85569a6c6ae))
+
+
+### Features
+
+* check-package-dependencies ([#849](https://github.com/christophehurpeau/pob/issues/849)) ([213969e](https://github.com/christophehurpeau/pob/commit/213969ed1ee4bc99ce431186a9d590a42d507d5e))
+
+
+
+
+
 # [7.9.0](https://github.com/christophehurpeau/pob/compare/pob@7.8.1...pob@7.9.0) (2021-02-14)